Main refactoring mostly done
This commit doesn't work though due to a circular dependency issue with the
board, move, and piece files.

At this point the Board object, which is supposed to replace the ChessEngine file
and the GameState class doesn't have all the logic it needs, and the pieces
themselves are not fully fleshed out yet.  This is an intermediary commit.