Updated move generation for pieces to ensure king can't make invalid moves
This commit includes a bunch of changes to make it possible for a king to filter
its possible moves by whether a certain move could place it into check.  To
support this logic, methods for determining if a piece could attack a given
square had to be added.

Additionally, this includes a quick performance test to compare the old move
generation logic with the new generation logic to see if there have been
improvements or not.