fixing bugs in move validation logic
added more tests to cover calls to the `can_attack_square` method for each of
the pieces in order to make sure that the logic for the method is working.
Since a lot of the move generation logic depends on this method being correct
having tests for it will hopefully make things easier long term.